Student Paper as part of an University project.
On this Paper we explain how the EM-Algorith works, and the considerations to be made.
In retrospective, we conclude that give the sparsity and amount of data needed to solve this State Space identifacion problem, better sampling and sparse solutions should be implemented.  
Being students we aknowledge we made several mistakes and for this reason the the algorithm didnt converge to rigth values.